create discount code and create cart order

+++ product/models.py
@@ -22,14 +22,14 @@ class Product(models.Model):
     title = models.CharField(max_length=50)
     description = models.TextField()
     price = models.IntegerField()
-    discount = models.SmallIntegerField()
+    discount = models.IntegerField()
     image = models.ImageField(upload_to='products')
     size = models.ManyToManyField(Size ,blank=True , null=True , related_name='products')
     color = models.ManyToManyField(Color , related_name='color')
 
     def __str__(self):
         return self.title
-    
+   
     
     
 class information(models.Model):
